modifica al tipo di availableBalance in Account per creare un conflitto

+++ src/code/Account.java
@@ -6,7 +6,7 @@ public class Account
 {
    private int accountNumber; // account number
    private int pin; // PIN for authentication
-   private double availableBalance; // funds available for withdrawal
+   private int availableBalance; // funds available for withdrawal
    private double totalBalance; // funds available + pending deposits
 
    // Account constructor initializes attributes
